moved file locations and references
