+++ 404.html
@@ -5,7 +5,7 @@
     <meta name="template" content="404"/>
     <meta name="404" content="local"/>
     <meta name="viewport" content="width=device-width, initial-scale=1"/>
-    <script src="/scripts/scripts.js" type="module"></script>
+    <script src="/events/scripts/scripts.js" type="module"></script>
     <link rel="icon" href="data:,">
     </head>
   <body>
