+++ head.html
@@ -1,7 +1,7 @@
 <!-- Modifying this file will impact your performance -->
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
-<script src="/scripts/fallback.js" nomodule></script>
-<script src="/scripts/scripts.js" type="module"></script>
-<link rel="stylesheet" href="/styles/styles.css">
+<script src="/events/scripts/fallback.js" nomodule></script>
+<script src="/events/scripts/scripts.js" type="module"></script>
+<link rel="stylesheet" href="/events/styles/styles.css">
 <style>body { display: none; }</style>
 <link rel="icon" href="data:,">
