+++ test/scripts/utils.test.js
@@ -1,6 +1,6 @@
 import { readFile } from '@web/test-runner-commands';
 import { expect } from '@esm-bundle/chai';
-import { decorateArea } from '../../scripts/utils.js';
+import { decorateArea } from '../../events/scripts/utils.js';
 
 document.head.innerHTML = await readFile({ path: './mocks/head.html' });
 const marqueeMain = await readFile({ path: './mocks/body-with-marquee.html' });
